requirements.txt
async-timeout==4.0.3
bjoern==3.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
dramatiq==1.17.0
dramatiq-dashboard==0.4.0
extend-me==1.1.5
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
odoo-rpc-client==1.2.0
prometheus_client==0.20.0
python-dotenv==1.0.1
redis==4.6.0
requests==2.31.0
simplejson==3.19.2
six==1.16.0
urllib3==2.2.1
